task-1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,10 +13,10 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	int i;
 	
 	if (array == NULL || action == NULL)
 		return;
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int) size; i++)
 		action(array[i]);
 }
